% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_marcxml.R
\name{read_marcxml}
\alias{read_marcxml}
\title{Read MARCXML into a tidy dataframe}
\usage{
read_marcxml(x, subfield_style = "hidden", repeating = "flat")
}
\arguments{
\item{x}{A file path string or a string of XML.}

\item{subfield_style}{Whether subfields are \code{hidden} (default), displayed \code{inline} (e.g., \verb{$a The Myth of Sysiphus}), or \code{unnested} into distinct columns. Only \code{hidden} is implemented.}

\item{repeating}{Whether repeating fields (e.g., subjects) are structured as \code{flat} strings (default), \code{list} objects, or \code{unnested} into multiple columns. Only \code{flat} is implemented.}
}
\value{
A tibble.
}
\description{
Read MARCXML into a tidy dataframe
}
\examples{
# TODO
}
