% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_names.R
\name{tag_names}
\alias{tag_names}
\title{Apply readable names to a MARC dataframe.}
\usage{
tag_names(df, style = "short")
}
\arguments{
\item{df}{A dataframe or dataframe extension (e.g., tibble)}

\item{style}{Whether names should be \code{short} (default), i.e., abbreviated, or \code{long}, using the official Library of Congress designations.}
}
\value{
A tibble.
}
\description{
Apply readable names to a MARC dataframe.
}
\examples{
# TODO
}
